bpm-ai-core/pyproject.toml
[tool.poetry]
name = "bpm-ai-core"
version = "1.2.0"
description = "Core AI abstractions and helpers."
authors = ["Bennet Krause <bennet.krause@holisticon.de>"]
repository = "https://github.com/holunda-io/bpm-ai"
homepage = "https://www.holisticon.de/"
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.3"
tenacity = "^8.2.3"
pydantic = "^2.6.4"
pillow = "^10.1.0"
pdf2image = "^1.17.0"
requests = "^2.31.0"
xmltodict = "^0.13.0"
aiohttp = "^3.9.3"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.23.5"
pytest-dotenv = "^0.5.2"

[tool.poetry.group.dev.dependencies]
openai = "^1.11.0"
anthropic = "^0.18.1"
mistralai = "^0.1.3"
langfuse = "^2.7.6"
setuptools = "^68.2.2"
faster-whisper = "^0.10.0"
lingua-language-detector = "^2.0.2"
pytesseract = "^0.3.10"
amazon-textract-prettyprinter = "^0.1.9"
aiobotocore = "^2.12.1"
azure-storage-blob = "^12.19.1"
azure-ai-translation-text = "^1.0.0b1"
azure-ai-documentintelligence = "^1.0.0b2"
torch = [
    { version = "=2.2.0", source="pypi", markers = "sys_platform == 'darwin'" },
    { version = "=2.2.0+cpu", source = "torch-cpu", markers = "sys_platform != 'darwin'" },
]
transformers = "^4.37.2"
sacremoses = "^0.1.1"
sentencepiece = "^0.1.99"
nltk = "^3.8.0"
spacy = [
    { version = "=3.7.2", markers = "sys_platform != 'darwin'" },
    { version = "=3.7.2", extras = ["apple"], markers = "sys_platform == 'darwin' and platform_machine == 'arm64'" },
]

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
